transcription/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-RESTful==0.3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pytz==2024.2
six==1.16.0
werkzeug==3.0.6
zipp==3.20.2
Pillow==11.0.0
Requests==2.32.3
supervision==0.24.0
torch==2.5.1
transformers==4.46.1
einops
timm
